+++ src/app/shared/shared.module.ts
@@ -1,16 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MouseEnterBackroundDirective } from './mouse-enter-backround.directive';
+import { ClickBorderColorDirective } from './directives/click-border-color.directive';
 
 
 
 @NgModule({
-  declarations: [MouseEnterBackroundDirective],
+  declarations: [MouseEnterBackroundDirective, ClickBorderColorDirective],
   imports: [
     CommonModule
   ],
   exports: [
-    MouseEnterBackroundDirective
+    MouseEnterBackroundDirective,
+    ClickBorderColorDirective
   ]
 })
 export class SharedModule { }